+++ tests/test-0741ui_locale_zh_SG.rb
@@ -1,0 +1,4 @@
+#!/usr/bin/ruby -w
+
+# T_741ui_locale_zh_SG
+test_ui_locale "zh_SG"